requirements.txt
numpy>=1.24
pandas>=2.1
pyyaml>=6.0
matplotlib>=3.8
seaborn>=0.13
pyarrow>=14.0
plotly>=5.22
kaleido>=0.2.1
requests>=2.31
streamlit>=1.36
python-dateutil>=2.8
